+++ app/src/main/java/com/konkuk/boost/data/course/RegistrationStatusData.kt
@@ -1,6 +1,6 @@
 package com.konkuk.boost.data.course
 
-import com.konkuk.boost.persistence.LikeCourseEntity
+import com.konkuk.boost.persistence.like.LikeCourseEntity
 
 data class RegistrationStatusData(
     val likeCourseEntity: LikeCourseEntity,
