+++ app/src/main/java/com/konkuk/boost/persistence/grade/GradeContract.kt
@@ -1,4 +1,4 @@
-package com.konkuk.boost.persistence
+package com.konkuk.boost.persistence.grade
 
 import android.provider.BaseColumns
