+++ app/src/main/java/com/konkuk/boost/persistence/simul/GraduationSimulationContract.kt
@@ -1,4 +1,4 @@
-package com.konkuk.boost.persistence
+package com.konkuk.boost.persistence.simul
 
 import android.provider.BaseColumns
