+++ app/src/main/java/com/konkuk/boost/persistence/simul/GraduationSimulationDao.kt
@@ -1,13 +1,13 @@
-package com.konkuk.boost.persistence
+package com.konkuk.boost.persistence.simul
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.OnConflictStrategy
 import androidx.room.Query
-import com.konkuk.boost.persistence.GraduationSimulationContract.GraduationSimulationEntry.ACQUIRED
-import com.konkuk.boost.persistence.GraduationSimulationContract.GraduationSimulationEntry.STANDARD
-import com.konkuk.boost.persistence.GraduationSimulationContract.GraduationSimulationEntry.TABLE_NAME
-import com.konkuk.boost.persistence.GraduationSimulationContract.GraduationSimulationEntry.USERNAME
+import com.konkuk.boost.persistence.simul.GraduationSimulationContract.GraduationSimulationEntry.ACQUIRED
+import com.konkuk.boost.persistence.simul.GraduationSimulationContract.GraduationSimulationEntry.STANDARD
+import com.konkuk.boost.persistence.simul.GraduationSimulationContract.GraduationSimulationEntry.TABLE_NAME
+import com.konkuk.boost.persistence.simul.GraduationSimulationContract.GraduationSimulationEntry.USERNAME
 
 @Dao
 interface GraduationSimulationDao {
