+++ app/src/main/java/com/konkuk/boost/repositories/GradeRepository.kt
@@ -3,10 +3,10 @@ package com.konkuk.boost.repositories
 import com.konkuk.boost.data.grade.GraduationSimulationResponse
 import com.konkuk.boost.data.grade.SubjectAreaCount
 import com.konkuk.boost.data.grade.UserInformationResponse
-import com.konkuk.boost.persistence.GradeEntity
-import com.konkuk.boost.persistence.GraduationSimulationEntity
-import com.konkuk.boost.persistence.RankEntity
-import com.konkuk.boost.persistence.SubjectAreaEntity
+import com.konkuk.boost.persistence.grade.GradeEntity
+import com.konkuk.boost.persistence.simul.GraduationSimulationEntity
+import com.konkuk.boost.persistence.rank.RankEntity
+import com.konkuk.boost.persistence.area.SubjectAreaEntity
 import com.konkuk.boost.utils.UseCase
 
 interface GradeRepository {
