+++ app/src/main/java/com/konkuk/boost/viewmodels/GraduationSimulationDetailViewModel.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.konkuk.boost.persistence.GradeEntity
+import com.konkuk.boost.persistence.grade.GradeEntity
 import com.konkuk.boost.repositories.GradeRepository
 import com.konkuk.boost.utils.UseCase
 import kotlinx.coroutines.Dispatchers
